requirements.txt
fastapi
diffusers
torch
torchvision
torchaudio
python-dotenv
uvicorn
accelerate